+++ include/select_menu.h
@@ -11,21 +11,21 @@
 class Select_Menu
 {
 public:
-    Select_Menu(bool enable_cancel, int nMenu_type, int nStartX, int nStartY);
+    Select_Menu(bool enable_cancel, u8 nMenu_type, int nStartX, int nStartY);
     int select_menu_main();
     void hide_menu();
     void show_menu();
     void clear_options();
-    void add_option(const byte *option, int return_value);
-    void set_lang(int nLang);
+    void add_option(const byte *option, u8 return_value);
+    void set_lang(u8 nLang);
 
 private:
     ptgb::vector<const byte*> menu_options;
-    ptgb::vector<int> return_values;
-    unsigned int curr_selection;
+    ptgb::vector<u8> return_values;
+    u16 curr_selection;
     bool cancel_enabled;
-    int menu_type;
-    int lang;
+    u8 menu_type;
+    u8 lang;
     int startTileX;
     int startTileY;
 };
